# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Parmana API
  version: 1.0.0
  description: >
    Parmana is an Execution Trust Infrastructure that ensures there is no gap
    between what humans decide and what AI systems do. The API enables creation,
    execution, verification, replay, and auditing of Business Transactions
    through cryptographically verifiable Execution Trust Records.


    **Every route requires a caller bearer key, except GET /health.** Send
    `Authorization: Bearer <key>` on every other request. Keys are issued by
    `scripts/generate-api-key.ts` and configured server-side via
    `PARMANA_API_KEYS`; only a hash of each key is ever held by the server,
    verified in constant time. A missing or invalid credential returns 401
    before a Business Transaction is even constructed, independent of Policy
    evaluation and gateway attestation, see `packages/api/src/middleware/caller-auth.ts`
    and [Authentication](/api-reference/authentication). Local development may
    set `PARMANA_AUTH_DISABLED=true` to skip this middleware entirely; that flag
    must never be set in a real deployment.

  contact:
    name: Parmana
    email: support@parmana.ai

  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: http://localhost:3000
    description: Local (packages/api, PORT env var, default 3000)
security:
  - bearerAuth: []
...
paths:
...
  /trust-records/{businessTransactionId}:
    get:
      tags:
        - Trust Records
      summary: Get Execution Trust Record
      operationId: getTrustRecord
      description: >
        Returns the complete Execution Trust Record for businessTransactionId
        (looked up by Business Transaction ID, despite the path segment name;
        there is no separate trustRecordId lookup).
      parameters:
        - name: businessTransactionId
          in: path
          required: true
          description: Business Transaction identifier (not the Trust Record's own trustRecordId).
          schema:
            type: string
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          description: Execution Trust Record found.
          content:
            application/json:
              schema:
                $ref: '../schemas/responses/execution-trust-record-response.schema.json'
              examples:
                vendor-payment-flow:
                  summary: Real captured response
                  value:
                    trustRecordId: 65fee934-532c-4feb-88d9-72c0cd912a81
                    businessTransactionId: 44b34a79-e0f2-49d7-a48e-e52fff88182e
                    transaction:
                      businessTransactionId: 44b34a79-e0f2-49d7-a48e-e52fff88182e
                      metadata:
                        businessTransactionId: 44b34a79-e0f2-49d7-a48e-e52fff88182e
                        correlationId: cc1ca975-4935-4e9a-a591-8897af6b56fb
                        sourceSystem: vendor-payment-service
                        submittedBy: ap-automation
                        submittedAt: '2026-07-13T17:36:00.835Z'
                      authority:
                        authorityId: db5c5260-bb4c-40a2-a081-2928b5fe7720
                        authorityType: SERVICE
                        principalId: ap-automation-svc
                        displayName: Accounts Payable Automation
                        issuedAt: '2026-07-13T17:36:00.835Z'
                      authorization:
                        authorizationId: 63bdc429-14e4-4ef4-8007-f3c0c9564261
                        authorityId: db5c5260-bb4c-40a2-a081-2928b5fe7720
                        purpose: Authorize vendor payment disbursement
                        issuedAt: '2026-07-13T17:36:00.835Z'
                      intent:
                        intentId: 1e77d0b3-2e57-4e31-b780-03abc92f1b69
                        authorizationId: 63bdc429-14e4-4ef4-8007-f3c0c9564261
                        action: payments:execute
                        target: vendor/V-500
                        parameters:
                          amount: 5000
                          currency: USD
                        createdAt: '2026-07-13T17:36:00.835Z'
                      policy:
                        name: vendor-payment
                        version: 2.0.0
                        schemaVersion: 1.0.0
                      signals:
                        vendorVerified: true
                        invoiceVerified: true
                        paymentApproved: true
                        sufficientFunds: true
                        paymentAmount: 5000
                        riskScore: 12
                      status: RECEIVED
                      createdAt: '2026-07-13T17:36:00.965Z'
                    overrides: []
                    executions:
                      - executionId: 90746492-7721-4228-b244-67bf8a45ca68
                        businessTransactionId: 44b34a79-e0f2-49d7-a48e-e52fff88182e
                        decision:
                          decisionId: 8ec880cb-677a-4889-96f1-5a18a4edc1b5
                          intentId: 1e77d0b3-2e57-4e31-b780-03abc92f1b69
                          policy:
                            name: vendor-payment
                            version: 2.0.0
                            schemaVersion: 1.0.0
                          signals:
                            vendorVerified: true
                            invoiceVerified: true
                            paymentApproved: true
                            sufficientFunds: true
                            paymentAmount: 5000
                            riskScore: 12
                          outcome: APPROVED
                          reason: Vendor payment authorized. Vendor verification, invoice verification, payment approval, funding, and risk assessment requirements were satisfied.
                          evaluatedAt: '2026-07-13T17:36:00.967Z'
                        status: COMPLETED
                        mode: SYNC
                        startedAt: '2026-07-13T17:36:00.968Z'
                        metadata:
                          authorizationId: dbc09266-2078-4d4e-93d0-f355eb7eebcc
                        evidence:
                          businessTransactionId: 44b34a79-e0f2-49d7-a48e-e52fff88182e
                          action: payments:execute
                          target: vendor/V-500
                          parameters:
                            amount: 5000
                            currency: USD
                          success: true
                          executedAt: '2026-07-13T17:36:00.970Z'
                          attributes:
                            connector:
                              connectorId: vendor-payment
                              connectorVersion: 1.0.0
                              capability: payments:execute
                              sanitizedEndpoint: vendor/V-500
                              credentialProviderId: environment
                              requestSummary:
                                businessTransactionId: 44b34a79-e0f2-49d7-a48e-e52fff88182e
                                action: payments:execute
                                target: vendor/V-500
                                parameters:
                                  amount: 5000
                                  currency: USD
                              responseSummary:
                                success: true
                                metadata: {}
                              startedAt: '2026-07-13T17:36:00.970Z'
                              completedAt: '2026-07-13T17:36:00.970Z'
                              connectorEvidenceHash: c97411026a82b608d9b0f8b137a64e26d8af24bdda09310b639e527e08a9aada
                        completedAt: '2026-07-13T17:36:00.970Z'
                    verifications:
                      - verificationId: e22354e0-7f22-49b2-bd2f-09e47881d788
                        businessTransactionId: 44b34a79-e0f2-49d7-a48e-e52fff88182e
                        status: VERIFIED
                        message: Execution Trust Record verified successfully.
                        verifiedAt: '2026-07-13T17:36:00.973Z'
                        trustRecordHash: e20a8861864f1a0a6c5fe00e64abca04de18a7d165a82ef9d7beb458c2096b3f
                    receipts:
                      - receiptId: 863947e6-9d36-492e-b9c9-78b0f3e9b6df
                        businessTransactionId: 44b34a79-e0f2-49d7-a48e-e52fff88182e
                        trustRecordHash: e20a8861864f1a0a6c5fe00e64abca04de18a7d165a82ef9d7beb458c2096b3f
                        receiptHash: c9666b08b42bb5ff77650b544c3c2183609f25e4d8dc09359b71a970c57ed3ce
                        issuedAt: '2026-07-13T17:36:00.973Z'
                        algorithm: ed25519
                        signature: Pcr42EvbaUbLogvVhYiV3g7uuwvZr82HUJhVptnXRxmTtiEGiz2SCdz5Z80xRDvmvSxpOSG5OnvMAaRzJXncBg==
                    createdAt: '2026-07-13T17:36:00.971Z'
                    updatedAt: '2026-07-13T17:36:00.971Z'
                    trustRecordHash: e20a8861864f1a0a6c5fe00e64abca04de18a7d165a82ef9d7beb458c2096b3f
                    signature:
                      algorithm: ed25519
                      keyId: default
                      value: gjDYmHUBKIrv7bT3uwAkt3wHeoSwz5uANXQV5MfsFgbUIh83KiufZwxht3F3Een+7K/ZW9SRn0F0J7ACJRLbAg==
                      signedAt: '2026-07-13T17:36:00.972Z'

        '404':
          description: Execution Trust Record not found.
          content:
            application/json:
              schema:
                $ref: '../schemas/common/error.schema.json'
              examples:
                notFound:
                  summary: Real captured response
                  value:
                    error: Execution Trust Record not found.
...
  /webhooks/razorpay:
    post:
      tags:
        - Webhooks
      summary: Receive a Razorpay webhook delivery
      operationId: receiveRazorpayWebhook
      description: >
        Verifies the HMAC-SHA256 signature (X-Razorpay-Signature) over the raw
        request body against RAZORPAY_WEBHOOK_SECRET, timing-safe. A valid,
        fresh event (deduplicated on X-Razorpay-Event-Id) is persisted to a
        pending-events store and acknowledged; nothing is processed inline —
        settlement/lifecycle processing is a future milestone (M4b, see
        docs/CLAIMS.md). Not registered at all (404) when
        RAZORPAY_WEBHOOK_SECRET is not configured.
      parameters:
        - name: X-Razorpay-Signature
          in: header
          required: true
          schema:
            type: string
          description: hex(HMAC-SHA256(rawBody, RAZORPAY_WEBHOOK_SECRET))
        - name: X-Razorpay-Event-Id
          in: header
          required: true
          schema:
            type: string
          description: >-
            Required once the signature has verified; a request missing it is
            rejected even with a valid signature.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '../schemas/requests/razorpay-webhook-event.schema.json'
      responses:
        '200':
          description: Verified event accepted (fresh) or acknowledged (duplicate).
          content:
            application/json:
              schema:
                $ref: '../schemas/responses/razorpay-webhook-response.schema.json'
              examples:
                accepted:
                  summary: Fresh event, persisted
                  value:
                    status: accepted
                acknowledged:
                  summary: Replayed event id, not reprocessed
                  value:
                    status: acknowledged
        '401':
          description: >-
            Missing/invalid signature, or a validly-signed request missing
            X-Razorpay-Event-Id.
          content:
            application/json:
              schema:
                $ref: '../schemas/common/error.schema.json'
              examples:
                invalidSignature:
                  summary: Bad signature
                  value:
                    error: Invalid signature.
        '404':
          description: Route not registered — RAZORPAY_WEBHOOK_SECRET is not configured.
        '413':
          description: Request body exceeded the size cap.
          content:
            application/json:
              schema:
                $ref: '../schemas/common/error.schema.json'
              examples:
                tooLarge:
                  summary: Oversized body
                  value:
                    error: Payload too large.
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        Caller API key issued by scripts/generate-api-key.ts. Sent as
        Authorization: Bearer <key>. Verified against a stored SHA-256 hash
        in constant time by packages/api/src/auth/StaticKeyAuthenticator.ts.
        Required on every route except GET /health. See
        /api-reference/authentication.
  responses:
    Unauthorized:
      description: >-
        Missing or invalid caller credential (StaticKeyAuthenticator returned
        no identity). Real captured response, packages/api/src/middleware/caller-auth.ts.
      content:
        application/json:
          schema:
            $ref: '../schemas/common/error.schema.json'
          examples:
            authRequired:
              summary: Real captured response, missing or invalid Authorization header
              value:
                error: authentication required